requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
starlette==0.27.0
pydantic==2.3.0
email-validator==2.0.0
python-jose[cryptography]==3.3.0
python-dotenv==0.19.1
cryptography==41.0.3
requests==2.31.0
schedule==1.2.0
qrcode==7.4.2
pillow==10.0.0
tabulate==0.9.0
pyngrok==6.0.0
python-telegram-bot==20.3
tenacity==8.2.2
openai==1.3.5
transformers==4.35.0
torch==2.0.1
telegram